requirements.txt
aiohttp
SQLAlchemy
pydantic
python-dotenv